Refactoring to avoid import cycle
