+++ code/go/internal/errors/errors_test.go
@@ -1,4 +1,4 @@
-package validator
+package errors
 
 import (
 	"errors"
