+++ code/go/pkg/validator/errors.go
@@ -1,8 +1,8 @@
 package validator
 
 import (
-	"github.com/elastic/package-spec/code/go/internal/validator"
+	"github.com/elastic/package-spec/code/go/internal/errors"
 )
 
 // ValidationErrors is an Error that contains a iterable collection of validation error messages.
-type ValidationErrors validator.ValidationErrors
+type ValidationErrors errors.ValidationErrors
